Update support for Discoverer v2
