+++ .editorconfig
@@ -10,7 +10,7 @@ indent_style = space
 insert_final_newline = true
 trim_trailing_whitespace = true
 
-[{*.yml}]
+[*.yml]
 indent_size = 2
 indent_style = space
