+++ _config/adaptors.yml
@@ -1,0 +1,19 @@
+---
+Name: discoverer-elastic-adaptors
+After: discoverer-adaptors
+Only:
+    envvarset: 'ENTERPRISE_SEARCH_API_SEARCH_KEY'
+---
+SilverStripe\Core\Injector\Injector:
+  SilverStripe\Discoverer\Query\Facet\FacetAdaptor:
+    class: SilverStripe\DiscovererElasticEnterprise\Query\Facet\FacetAdaptor
+  SilverStripe\Discoverer\Query\Filter\CriteriaAdaptor:
+    class: SilverStripe\DiscovererElasticEnterprise\Query\Filter\CriteriaAdaptor
+  SilverStripe\Discoverer\Query\Filter\CriterionAdaptor:
+    class: SilverStripe\DiscovererElasticEnterprise\Query\Filter\CriterionAdaptor
+  SilverStripe\Discoverer\Service\Interfaces\ProcessAnalyticsAdaptor:
+    class: SilverStripe\DiscovererElasticEnterprise\Service\Adaptors\ProcessAnalyticsAdaptor
+  SilverStripe\Discoverer\Service\Interfaces\QuerySuggestionAdaptor:
+    class: SilverStripe\DiscovererElasticEnterprise\Service\Adaptors\QuerySuggestionAdaptor
+  SilverStripe\Discoverer\Service\Interfaces\SearchAdaptor:
+    class: SilverStripe\DiscovererElasticEnterprise\Service\Adaptors\SearchAdaptor
