--- _config/config.yml
@@ -1,20 +1,0 @@
----
-Name: search-elastic-config
-Only:
-    envvarset: 'ENTERPRISE_SEARCH_API_SEARCH_KEY'
----
-SilverStripe\Core\Injector\Injector:
-  SilverStripe\Discoverer\Query\Facet\FacetAdaptor:
-    class: SilverStripe\DiscovererElasticEnterprise\Query\Facet\FacetAdaptor
-  SilverStripe\Discoverer\Query\Filter\CriteriaAdaptor:
-    class: SilverStripe\DiscovererElasticEnterprise\Query\Filter\CriteriaAdaptor
-  SilverStripe\Discoverer\Query\Filter\CriterionAdaptor:
-    class: SilverStripe\DiscovererElasticEnterprise\Query\Filter\CriterionAdaptor
-  SilverStripe\Discoverer\Service\SearchServiceAdaptor:
-    class: SilverStripe\DiscovererElasticEnterprise\Service\SearchServiceAdaptor
-  Elastic\EnterpriseSearch\Client.searchClient:
-    factory: SilverStripe\DiscovererElasticEnterprise\Service\ClientFactory
-    constructor:
-      host: '`ENTERPRISE_SEARCH_ENDPOINT`'
-      token: '`ENTERPRISE_SEARCH_API_SEARCH_KEY`'
-      http_client: '%$GuzzleHttp\Client'
