+++ src/Service/ClientFactory.php
@@ -5,6 +5,7 @@ namespace SilverStripe\DiscovererElasticEnterprise\Service;
 use Elastic\EnterpriseSearch\Client;
 use Exception;
 use SilverStripe\Core\Injector\Factory;
+use SilverStripe\Core\Injector\Injector;
 
 class ClientFactory implements Factory
 {
@@ -47,7 +48,7 @@ class ClientFactory implements Factory
             'client' => $httpClient,
         ];
 
-        return new Client($config);
+        return Injector::inst()->create(Client::class, $config);
     }
 
 }
